requirements.txt
torch
torchvision
ftfy
h5py
regex
tqdm
Ninja
transformers
tensorboard
pytorch-fid
open-clip-torch
pycocotools
git+https://github.com/openai/CLIP.git
